Test api integrationtests

+++ backend/database/handler/databaseHandler.py
@@ -21,10 +21,13 @@ class DatabaseHandler:
     
     
     @staticmethod
-    def select(query: str, parameter: str = None) -> list[DatabaseObject]:
+    def select(query: str, parameter: tuple = None) -> list[DatabaseObject]:
         DatabaseHandler._CURSOR = DatabaseHandler._CONNECTION.cursor()
         if parameter == None:
             DatabaseHandler._CURSOR.execute(query)
+        elif type(parameter[0]) == str:
+            form = f"{parameter[0]}"
+            DatabaseHandler._CURSOR.execute(query, (form,))
         else:
             DatabaseHandler._CURSOR.execute(query, parameter)
         resultSet: list[DatabaseObject] = DatabaseHandler._CURSOR.fetchall()
@@ -77,21 +80,21 @@ class DatabaseHandler:
     
     #get all possible warning messages within the Database
     @staticmethod
-    def selectWarningMessages() -> list[str]:
-        resultSet: list[DatabaseObject] = DatabaseHandler.select(f"SELECT * FROM warning")
+    def selectWarningMessages() -> list[Warning]:
+        resultSet: list[Warning] = DatabaseHandler.select(f"SELECT * FROM warning")
         warningMessages: list[str] = []
         for result in resultSet:
             warning: Warning = Warning(DatabaseObject(result))
-            warningMessages.append(warning.getType())
+            warningMessages.append(warning)
         return warningMessages
     
     @staticmethod
-    def selectErrorMessages() -> list[str]:
+    def selectErrorMessages() -> list[Error]:
         resultSet: list[DatabaseObject] = DatabaseHandler.select(f"SELECT * FROM error")
         errorMessages: list[str] = []
         for result in resultSet:
             error: Error = Error(DatabaseObject(result))
-            errorMessages.append(error.getType())
+            errorMessages.append(error)
         return errorMessages
     
     @staticmethod
@@ -171,7 +174,9 @@ class DatabaseHandler:
     
     @staticmethod
     def deleteProgramState(id: int) -> None:
-        DatabaseHandler.delete("DELETE FROM program_state WHERE program_state_id = ?", (id,))
+        programState: ProgramState = DatabaseHandler.selectProgramState(id)
+        if(programState != None):
+            DatabaseHandler.delete("DELETE FROM program_state WHERE program_state_id = ?", (id,))
     
     @staticmethod
     def deleteMachineStateById(id: int) -> None:
